updating filternet examples
