+++ docs/examples/filter_components/model_templates/lgn_on_off_model.json
