+++ docs/examples/filter_graitings/network/lgn_node_types.csv
@@ -1,4 +1,4 @@
-node_type_id dynamics_params non_dom_params model_type model_template sf_sep ei
-100 tOFF_TF15_3.44215357_-2.11509939_8.27421573_20.0_0.0_ic.json  virtual lgnmodel:tOFF_TF15  e
-101 sOFF_TF4_3.5_-2.0_10.0_60.0_15.0_ic.json sON_TF4_3.5_-2.0_30.0_60.0_25.0_ic.json virtual lgnmodel:sONsOFF_001 6.0 e
-102 tOFF_TF8_4.222_-2.404_8.545_23.019_0.0_ic.json sON_TF4_3.5_-2.0_30.0_60.0_25.0_ic.json virtual lgnmodel:sONtOFF_001 4.0 e
+node_type_id dynamics_params sf_sep model_template model_type ei non_dom_params
+100 tOFF_TF15_3.44215357_-2.11509939_8.27421573_20.0_0.0_ic.json  lgnmodel:tOFF_TF15 virtual e 
+101 sOFF_TF4_3.5_-2.0_10.0_60.0_15.0_ic.json 6.0 lgnmodel:sONsOFF_001 virtual e sON_TF4_3.5_-2.0_30.0_60.0_25.0_ic.json
+102 tOFF_TF8_4.222_-2.404_8.545_23.019_0.0_ic.json 4.0 lgnmodel:sONtOFF_001 virtual e sON_TF4_3.5_-2.0_30.0_60.0_25.0_ic.json
