+++ docs/examples/filter_graitings_preset/build_network.py
@@ -5,8 +5,8 @@ import numpy as np
 from bmtk.builder import NetworkBuilder
 
 
-X_grids = 2 #15
-Y_grids = 2 #10
+X_grids = 2  # 15
+Y_grids = 2  # 10
 X_len = 240.0  # In linear degrees
 Y_len = 120.0  # In linear degrees
 
@@ -44,7 +44,6 @@ def get_filter_spatial_size(N, X_grids, Y_grids, size_range):
 
 
 def get_filter_temporal_params(N, X_grids, Y_grids, model):
-
     # Total number of cells
     N_total = N * X_grids * Y_grids
 
@@ -61,9 +60,7 @@ def get_filter_temporal_params(N, X_grids, Y_grids, model):
     tOFF_fn = os.path.join(basepath, 'tOFF_TF8_4.222_-2.404_8.545_23.019_0.0_ic.pkl')  # best chosen fit for tOFF 8 Hz
     sON_fn = os.path.join(basepath, 'sON_TF4_3.5_-2.0_30.0_60.0_25.0_ic.pkl')  # best chosen fit for sON 4 Hz
 
-    #print(open(sOFF_fn))
     sOFF_prs = pickle.load(open(sOFF_fn, 'rb')) #, encoding='latin1')
-    #print(sOFF_prs)
     tOFF_prs = pickle.load(open(tOFF_fn, 'rb'))
     sON_prs = pickle.load(open(sON_fn, 'rb'))
 
@@ -81,6 +78,7 @@ def get_filter_temporal_params(N, X_grids, Y_grids, model):
         wts = sOFF_prs['opt_wts']
         wts_dom_0 = np.random.uniform(lower_jitter * wts[0], upper_jitter * wts[0], N_total)
         wts_dom_1 = np.random.uniform(lower_jitter * wts[1], upper_jitter * wts[1], N_total)
+
         wts =  sON_prs['opt_wts']
         wts_non_dom_0 = np.random.uniform(lower_jitter * wts[0], upper_jitter * wts[0], N_total)
         wts_non_dom_1 = np.random.uniform(lower_jitter * wts[1], upper_jitter * wts[1], N_total)
@@ -88,6 +86,7 @@ def get_filter_temporal_params(N, X_grids, Y_grids, model):
         delays = sOFF_prs['opt_delays']
         delays_dom_0 = np.random.uniform(lower_jitter * delays[0], upper_jitter * delays[0], N_total)
         delays_dom_1 = np.random.uniform(lower_jitter * delays[1], upper_jitter * delays[1], N_total)
+
         delays = sON_prs['opt_delays']
         delays_non_dom_0 = np.random.uniform(lower_jitter * delays[0], upper_jitter * delays[0], N_total)
         delays_non_dom_1 = np.random.uniform(lower_jitter * delays[1], upper_jitter * delays[1], N_total)
@@ -97,10 +96,10 @@ def get_filter_temporal_params(N, X_grids, Y_grids, model):
         tuning_angles = np.random.uniform(0, 360., N_total)
 
     elif model == 'sONtOFF_001':
-
         kpeaks = tOFF_prs['opt_kpeaks']
         kpeaks_dom_0 = np.random.uniform(lower_jitter * kpeaks[0], upper_jitter * kpeaks[0], N_total)
         kpeaks_dom_1 = np.random.uniform(lower_jitter * kpeaks[1], upper_jitter * kpeaks[1], N_total)
+
         kpeaks = sON_prs['opt_kpeaks']
         kpeaks_non_dom_0 = np.random.uniform(lower_jitter * kpeaks[0], upper_jitter * kpeaks[0], N_total)
         kpeaks_non_dom_1 = np.random.uniform(lower_jitter * kpeaks[1], upper_jitter * kpeaks[1], N_total)
@@ -108,6 +107,7 @@ def get_filter_temporal_params(N, X_grids, Y_grids, model):
         wts = tOFF_prs['opt_wts']
         wts_dom_0 = np.random.uniform(lower_jitter * wts[0], upper_jitter * wts[0], N_total)
         wts_dom_1 = np.random.uniform(lower_jitter * wts[1], upper_jitter * wts[1], N_total)
+
         wts = sON_prs['opt_wts']
         wts_non_dom_0 = np.random.uniform(lower_jitter * wts[0], upper_jitter * wts[0], N_total)
         wts_non_dom_1 = np.random.uniform(lower_jitter * wts[1], upper_jitter * wts[1], N_total)
@@ -115,6 +115,7 @@ def get_filter_temporal_params(N, X_grids, Y_grids, model):
         delays = tOFF_prs['opt_delays']
         delays_dom_0 = np.random.uniform(lower_jitter * delays[0], upper_jitter * delays[0], N_total)
         delays_dom_1 = np.random.uniform(lower_jitter * delays[1], upper_jitter * delays[1], N_total)
+
         delays = sON_prs['opt_delays']
         delays_non_dom_0 = np.random.uniform(lower_jitter * delays[0], upper_jitter * delays[0], N_total)
         delays_non_dom_1 = np.random.uniform(lower_jitter * delays[1], upper_jitter * delays[1], N_total)
@@ -162,45 +163,46 @@ def get_filter_temporal_params(N, X_grids, Y_grids, model):
         sf_sep = np.nan * np.zeros(N_total)
         tuning_angles =  np.nan * np.zeros(N_total)
 
-    return np.column_stack((kpeaks_dom_0, kpeaks_dom_1, wts_dom_0, wts_dom_1, delays_dom_0, delays_dom_1,
-                            kpeaks_non_dom_0, kpeaks_non_dom_1, wts_non_dom_0, wts_non_dom_1,
-                            delays_non_dom_0, delays_non_dom_1, tuning_angles, sf_sep))
-
+    return np.column_stack(
+        (kpeaks_dom_0, kpeaks_dom_1, wts_dom_0, wts_dom_1, delays_dom_0, delays_dom_1,
+         kpeaks_non_dom_0, kpeaks_non_dom_1, wts_non_dom_0, wts_non_dom_1, delays_non_dom_0,
+         delays_non_dom_1, tuning_angles, sf_sep)
+    )
 
 
 lgn_models = [
     # {
-    #     'N': 7, 'ei': 'e', 'model_type': 'virtual', 'pop_name': 'sON_TF1', 'size_range': [2, 10]
+    #     'N': 7, 'ei': 'e', 'model_type': 'virtual', 'model_template': 'lgnmodel:sON_TF1', 'size_range': [2, 10]
     # },
     # {
-    #     'N': 5, 'ei': 'e', 'model_type': 'virtual', 'pop_name': 'sON_TF2', 'size_range': [2, 10]
+    #     'N': 5, 'ei': 'e', 'model_type': 'virtual', 'model_template': 'lgnmodel:sON_TF2', 'size_range': [2, 10]
     # },
     # {
-    #     'N': 7, 'ei': 'e', 'model_type': 'virtual', 'pop_name': 'sON_TF4', 'size_range': [2, 10]
+    #     'N': 7, 'ei': 'e', 'model_type': 'virtual', 'model_template': 'lgnmodel:sON_TF4', 'size_range': [2, 10]
     # },
     # {
-    #     'N': 15, 'ei': 'e', 'model_type': 'virtual', 'pop_name': 'sON_TF8', 'size_range': [2, 10]
+    #     'N': 15, 'ei': 'e', 'model_type': 'virtual', 'model_template': 'lgnmodel:sON_TF8', 'size_range': [2, 10]
     # },
     # {
-    #     'N': 8, 'ei': 'e', 'model_type': 'virtual', 'pop_name': 'sOFF_TF1', 'size_range': [2, 10]
+    #     'N': 8, 'ei': 'e', 'model_type': 'virtual', 'model_template': 'lgnmodel:sOFF_TF1', 'size_range': [2, 10]
     # },
     # {
-    #     'N': 8, 'ei': 'e', 'model_type': 'virtual', 'pop_name': 'sOFF_TF2', 'size_range': [2, 10]
+    #     'N': 8, 'ei': 'e', 'model_type': 'virtual', 'model_template': 'lgnmodel:sOFF_TF2', 'size_range': [2, 10]
     # },
     # {
-    #     'N': 15, 'ei': 'e', 'model_type': 'virtual', 'pop_name': 'sOFF_TF4', 'size_range': [2, 10]
+    #     'N': 15, 'ei': 'e', 'model_type': 'virtual', 'model_template': 'lgnmodel:sOFF_TF4', 'size_range': [2, 10]
     # },
     # {
-    #     'N': 8, 'ei': 'e', 'model_type': 'virtual', 'pop_name': 'sOFF_TF8', 'size_range': [2, 10]
+    #     'N': 8, 'ei': 'e', 'model_type': 'virtual', 'model_template': 'lgnmodel:sOFF_TF8', 'size_range': [2, 10]
     # },
     # {
-    #     'N': 7, 'ei': 'e', 'model_type': 'virtual', 'pop_name': 'sOFF_TF15', 'size_range': [2, 10]
+    #     'N': 7, 'ei': 'e', 'model_type': 'virtual', 'model_template': 'lgnmodel:sOFF_TF15', 'size_range': [2, 10]
     # },
     # {
-    #    'N': 10, 'ei': 'e', 'model_type': 'virtual', 'pop_name': 'tOFF_TF4', 'size_range': [2, 10]
+    #    'N': 10, 'ei': 'e', 'model_type': 'virtual', 'model_template': 'lgnmodel:tOFF_TF4', 'size_range': [2, 10]
     # },
     # {
-    #    'N': 5, 'ei': 'e', 'model_type': 'virtual', 'pop_name': 'tOFF_TF8', 'size_range': [2, 10]
+    #    'N': 5, 'ei': 'e', 'model_type': 'virtual', 'model_template': 'lgnmodel:tOFF_TF8', 'size_range': [2, 10]
     # },
     {
         'N': 8, 'ei': 'e', 'model_type': 'virtual', 'model_template': 'lgnmodel:tOFF_TF15', 'size_range': [2, 10]
@@ -213,9 +215,7 @@ lgn_models = [
     }
 ]
 
-LGN = NetworkBuilder('lgn')
-xcoords = []
-ycoords = []
+lgn = NetworkBuilder('lgn')
 for params in lgn_models:
     # Get position of lgn cells and keep track of the averaged location
     # For now, use randomly generated values
@@ -223,8 +223,6 @@ for params in lgn_models:
 
     # Get positional coordinates of cells
     positions = generate_positions_grids(params['N'], X_grids, Y_grids, X_len, Y_len)
-    xcoords += [p[0] for p in positions]
-    ycoords += [p[1] for p in positions]
 
     # Get spatial filter size of cells
     filter_sizes = get_filter_spatial_size(params['N'], X_grids, Y_grids, params['size_range'])
@@ -232,33 +230,35 @@ for params in lgn_models:
     # Get filter temporal parameters
     filter_params = get_filter_temporal_params(params['N'], X_grids, Y_grids, params['model_template'])
 
-    LGN.add_nodes(N=total_N, ei=params['ei'],
-                  model_type=params['model_type'],
-                  model_processing='preset_params',  # Used to indicate parameters are preset
-                  model_template=params['model_template'],
-                  x=positions[:, 0],
-                  y=positions[:, 1],
-                  spatial_size=filter_sizes,
-                  # kpeaks_dom_0=filter_params[:, 0],
-                  # kpeaks_dom_1=filter_params[:, 1],
-                  kpeaks=filter_params[:, 0:2],
-                  #weight_dom_0=filter_params[:, 2],
-                  #weight_dom_1=filter_params[:, 3],
-                  weights=filter_params[:, 2:4],
-                  #delay_dom_0=filter_params[:, 4],
-                  #delay_dom_1=filter_params[:, 5],
-                  delays=filter_params[:, 4:6],
+    lgn.add_nodes(
+        N=total_N,
+        ei=params['ei'],
+        model_type=params['model_type'],
+        model_processing='preset_params',  # Used to indicate parameters are preset
+        model_template=params['model_template'],
+        x=positions[:, 0],
+        y=positions[:, 1],
+        spatial_size=filter_sizes,
+        # kpeaks_dom_0=filter_params[:, 0],
+        # kpeaks_dom_1=filter_params[:, 1],
+        kpeaks=filter_params[:, 0:2],
+        #weight_dom_0=filter_params[:, 2],
+        #weight_dom_1=filter_params[:, 3],
+        weights=filter_params[:, 2:4],
+        #delay_dom_0=filter_params[:, 4],
+        #delay_dom_1=filter_params[:, 5],
+        delays=filter_params[:, 4:6],
 
-                  #kpeaks_non_dom_0=filter_params[:, 6],
-                  # kpeaks_non_dom_1=filter_params[:, 7],
-                  kpeaks_non_dom=filter_params[:, 6:8],
-                  weight_non_dom_0=filter_params[:, 8],
-                  weight_non_dom_1=filter_params[:, 9],
-                  delay_non_dom_0=filter_params[:, 10],
-                  delay_non_dom_1=filter_params[:, 11],
-                  tuning_angle=filter_params[:, 12],
-                  sf_sep=filter_params[:, 13],
-                  )
+        # kpeaks_non_dom_0=filter_params[:, 6],
+        # kpeaks_non_dom_1=filter_params[:, 7],
+        kpeaks_non_dom=filter_params[:, 6:8],
+        weight_non_dom_0=filter_params[:, 8],
+        weight_non_dom_1=filter_params[:, 9],
+        delay_non_dom_0=filter_params[:, 10],
+        delay_non_dom_1=filter_params[:, 11],
+        tuning_angle=filter_params[:, 12],
+        sf_sep=filter_params[:, 13],
+    )
 
-LGN.build()
-LGN.save(output_dir='network')
+lgn.build()
+lgn.save(output_dir='network')
