+++ docs/examples/filter_graitings_preset/config.circuit.json
@@ -1,0 +1,20 @@
+{
+  "manifest": {
+    "$BASE_DIR": ".",
+    "$NETWORK_DIR": "$BASE_DIR/network",
+    "$COMPONENT_DIR": "$BASE_DIR/../filter_components"
+  },
+
+  "components": {
+    "filter_models_dir": "$COMPONENT_DIR/model_templates"
+  },
+
+  "networks": {
+    "nodes": [
+      {
+        "nodes_file": "$NETWORK_DIR/lgn_nodes.h5",
+        "node_types_file": "$NETWORK_DIR/lgn_node_types.csv"
+      }
+    ]
+  }
+}
