+++ docs/examples/filter_graitings_preset/config.simulation.json
@@ -1,0 +1,47 @@
+{
+  "manifest": {
+    "$BASE_DIR": ".",
+    "$OUTPUT_DIR": "$BASE_DIR/output"
+  },
+
+  "run": {
+    "tstop": 2000.0,
+    "dt": 0.1,
+    "overwrite_output_dir": true
+  },
+
+  "target_simulator":"LGNModel",
+
+  "conditions": {
+    "celsius": 34.0,
+    "v_init": -80
+  },
+
+  "inputs": {
+    "LGN_spikes": {
+      "input_type": "movie",
+      "module": "graiting",
+      "row_size": 120,
+      "col_size": 240,
+      "gray_screen_dur": 0.5,
+      "cpd": 0.04,
+      "temporal_f": 4.0,
+      "contrast": 0.8,
+      "evaluation_options": {
+        "downsample": 1,
+        "separable": true
+      }
+    }
+  },
+
+  "output":{
+    "log_file": "log.txt",
+    "output_dir": "$OUTPUT_DIR",
+    "rates_csv": "rates.csv",
+    "spikes_csv": "spikes.csv",
+    "spikes_h5": "spikes.h5",
+    "overwrite_output_dir": true
+  },
+
+  "network": "config.circuit.json"
+}
