+++ docs/examples/filter_looming/config.json
@@ -1,24 +1,23 @@
 {
-    "manifest": {
-        "$BASE_DIR": "${configdir}",
-        "$OUTPUT_DIR": "$BASE_DIR/output",
-        "$INPUT_DIR": "$BASE_DIR/movie_files",
-        "$NETWORK_DIR": "$BASE_DIR/network",
-        "$COMPONENT_DIR": "$BASE_DIR/../filter_components"
-    },
+  "manifest": {
+    "$BASE_DIR": "${configdir}",
+    "$OUTPUT_DIR": "$BASE_DIR/output",
+    "$INPUT_DIR": "$BASE_DIR/movie_files",
+    "$NETWORK_DIR": "$BASE_DIR/network",
+    "$COMPONENT_DIR": "$BASE_DIR/../filter_components"
+  },
 
-    "run": {
-        "tstop": 2050.0,
-        "dt": 0.1,
-        "overwrite_output_dir": true
-    },
+  "run": {
+    "tstop": 2050.0,
+    "dt": 0.1
+  },
 
-    "target_simulator": "LGNModel",
+  "target_simulator": "LGNModel",
 
-    "conditions": {
-        "jitter_lower": 1.0,
-        "jitter_upper": 1.0
-    },
+  "conditions": {
+    "jitter_lower": 1.0,
+    "jitter_upper": 1.0
+  },
 
   "inputs": {
     "LGN_spikes": {
@@ -32,26 +31,25 @@
     }
   },
 
+  "output": {
+    "output_dir": "$OUTPUT_DIR",
+    "log_file": "$OUTPUT_DIR/log.txt",
+    "rates_csv": "$OUTPUT_DIR/rates.csv",
+    "spikes_csv": "$OUTPUT_DIR/spikes.csv",
+    "spikes_h5": "$OUTPUT_DIR/spikes.h5",
+    "overwrite_output_dir": true
+  },
 
-    "output": {
-        "log_file": "$OUTPUT_DIR/log.txt",
-        "output_dir": "$OUTPUT_DIR",
-        "rates_csv": "$OUTPUT_DIR/rates.csv",
-        "spikes_csv": "$OUTPUT_DIR/spikes.csv",
-        "spikes_h5": "$OUTPUT_DIR/spikes.h5",
-        "overwrite_output_dir": true
-    },
+  "components": {
+    "filter_models_dir": "$COMPONENT_DIR/model_templates"
+  },
 
-    "components": {
-        "filter_models_dir": "$COMPONENT_DIR/model_templates"
-    },
-
-    "networks": {
-        "nodes": [
-            {
-                "nodes_file": "$NETWORK_DIR/lgn_nodes.h5",
-                "node_types_file": "$NETWORK_DIR/lgn_node_types.csv"
-            }
-        ]
-    }
+  "networks": {
+    "nodes": [
+      {
+        "nodes_file": "$NETWORK_DIR/lgn_nodes.h5",
+        "node_types_file": "$NETWORK_DIR/lgn_node_types.csv"
+      }
+    ]
+  }
 }
