+++ docs/examples/filter_looming/config.simulation.json
@@ -1,0 +1,41 @@
+{
+  "manifest": {
+    "$BASE_DIR": ".",
+    "$OUTPUT_DIR": "$BASE_DIR/output"
+  },
+
+  "run": {
+    "tstop": 2050.0,
+    "dt": 0.1
+  },
+
+  "target_simulator": "LGNModel",
+
+  "conditions": {
+    "jitter_lower": 1.0,
+    "jitter_upper": 1.0
+  },
+
+  "inputs": {
+    "LGN_spikes": {
+      "input_type": "movie",
+      "module": "looming",
+      "row_size": 120,
+      "col_size": 240,
+      "frame_rate": 1000.0,
+      "gray_screen_dur": 0.5,
+      "t_looming": 1.0
+    }
+  },
+
+  "output": {
+    "output_dir": "$OUTPUT_DIR",
+    "log_file": "log.txt",
+    "rates_csv": "rates.csv",
+    "spikes_csv": "spikes.csv",
+    "spikes_h5": "spikes.h5",
+    "overwrite_output_dir": true
+  },
+
+  "network": "config.circuit.json"
+}
