+++ docs/examples/filter_looming/run_filternet.py
@@ -15,4 +15,4 @@ if __name__ == '__main__':
     if __file__ != sys.argv[-1]:
         run(sys.argv[-1])
     else:
-        run('config.json')
+        run('config.simulation.json')
