+++ docs/examples/filter_movie/config.circuit.json
@@ -1,0 +1,20 @@
+{
+  "manifest": {
+    "$BASE_DIR": ".",
+    "$COMPONETS_DIR": "$BASE_DIR/../filter_components",
+    "$NETWORK_DIR": "$BASE_DIR/network"
+  },
+
+  "components": {
+    "filter_models_dir": "$BASE_DIR"
+  },
+
+  "networks": {
+    "nodes": [
+      {
+        "nodes_file": "$NETWORK_DIR/lgn_nodes.h5",
+        "node_types_file": "$NETWORK_DIR/lgn_node_types.csv"
+      }
+    ]
+  }
+}
