--- docs/examples/filter_movie/config.json
@@ -1,56 +1,0 @@
-{
-    "manifest": {
-        "$BASE_DIR": "${configdir}",
-        "$OUTPUT_DIR": "$BASE_DIR/output",
-        "$INPUT_DIR": "$BASE_DIR/movie_files",
-        "$NETWORK_DIR": "$BASE_DIR/network"
-    },
-
-    "run": {
-        "tstop": 2050.0,
-        "dt": 0.1,
-        "overwrite_output_dir": true
-    },
-
-    "target_simulator": "NEURON",
-
-    "conditions": {
-        "jitter_lower": 1.0,
-        "jitter_upper": 1.0
-    },
-
-    "inputs": {
-        "movie_input": {
-            "input_type": "movie",
-            "module": "movie",
-            "data_file": "$INPUT_DIR/test_scene.npy",
-            "frame_rate": 30.0,
-            "evaluation_options": {
-                "t_min": 3.0,
-                "t_max": 4.0
-            }
-        }
-    },
-
-    "output": {
-        "log_file": "$OUTPUT_DIR/log.txt",
-        "output_dir": "$OUTPUT_DIR",
-        "rates_csv": "$OUTPUT_DIR/rates.csv",
-        "spikes_csv": "$OUTPUT_DIR/spikes.csv",
-        "spikes_h5": "$OUTPUT_DIR/spikes.h5",
-        "overwrite_output_dir": true
-    },
-
-    "components": {
-        "filter_models_dir": "$BASE_DIR"
-    },
-
-    "networks": {
-        "nodes": [
-            {
-                "nodes_file": "$NETWORK_DIR/lgn_nodes.h5",
-                "node_types_file": "$NETWORK_DIR/lgn_node_types.csv"
-            }
-        ]
-    }
-}
