+++ docs/examples/filter_movie/config.simulation.json
@@ -1,0 +1,43 @@
+{
+  "manifest": {
+    "$BASE_DIR": ".",
+    "$OUTPUT_DIR": "$BASE_DIR/output",
+    "$INPUT_DIR": "$BASE_DIR/movie_files"
+  },
+
+  "run": {
+    "tstop": 2050.0,
+    "dt": 0.1
+  },
+
+  "target_simulator": "LGNModel",
+
+  "conditions": {
+    "jitter_lower": 1.0,
+    "jitter_upper": 1.0
+  },
+
+  "inputs": {
+    "movie_input": {
+      "input_type": "movie",
+      "module": "movie",
+      "data_file": "$INPUT_DIR/test_scene.npy",
+      "frame_rate": 1000.0,
+      "evaluation_options": {
+        "t_min": 3.0,
+        "t_max": 4.0
+      }
+    }
+  },
+
+  "output": {
+    "output_dir": "$OUTPUT_DIR",
+    "log_file": "log.txt",
+    "rates_csv": "rates.csv",
+    "spikes_csv": "spikes.csv",
+    "spikes_h5": "spikes.h5",
+    "overwrite_output_dir": true
+  },
+
+  "network": "config.circuit.json"
+}
